cheri: drivers: pci: Handle capabilities in drivers/pci/
- For dynamic IDs allow a uintptr_t as the driver data, too.
- Allow the user to pass in an address via new_id_store and
  derive provenance from the existing ID table in case of a match.
- Annotate uintptr_t casts.

+++ drivers/pci/controller/pcie-rzg3s-host.c
@@ -189,7 +189,7 @@ struct rzg3s_pcie_msi {
 	DECLARE_BITMAP(map, RZG3S_PCI_MSI_INT_NR);
 	dma_addr_t dma_addr;
 	dma_addr_t window_base;
-	unsigned long pages;
+	uintptr_t pages;
 	struct mutex map_lock;
 	int irq;
 };
